Major refactor on uniswap-forks masterchef.
