+++ subgraphs/uniswap-forks/protocols/apeswap/src/common/handlers/handleRewardMini.ts
@@ -9,7 +9,7 @@ import {
 import { INT_ZERO, MasterChef } from "../../../../../src/common/constants";
 import { getOrCreateToken } from "../../../../../src/common/getters";
 import { getRewardsPerDay } from "../../../../../src/common/rewards";
-import { getOrCreateMasterChef } from "../helpers";
+import { getOrCreateMasterChef } from "../../../../../src/common/masterchef/helpers";
 import {
   convertTokenToDecimal,
   roundToWholeNumber,
