+++ subgraphs/uniswap-forks/protocols/apeswap/src/mappings/masterchef/rewardMini.ts
@@ -19,7 +19,7 @@ import {
   createMasterChefStakingPool,
   getOrCreateMasterChef,
   updateMasterChefTotalAllocation,
-} from "../../common/helpers";
+} from "../../../../../src/common/masterchef/helpers";
 
 // A deposit or stake for the pool specific MasterChef.
 export function handleDeposit(event: Deposit): void {
