+++ subgraphs/uniswap-forks/protocols/trader-joe/src/common/constants.ts
@@ -1,3 +1,5 @@
+import { BigInt } from "@graphprotocol/graph-ts";
+
 ////////////////////
 //////Versions//////
 ////////////////////
@@ -6,3 +8,6 @@ export const PROTOCOL_SUBGRAPH_VERSION = "1.0.1";
 export const PROTOCOL_METHODOLOGY_VERSION = "1.0.0";
 export const PROTOCOL_NAME = "Trader Joe";
 export const PROTOCOL_SLUG = "trader-joe";
+
+export const TRADER_JOE_AVALANCHE_REWARD_TOKEN_RATE =
+  BigInt.fromI32(30000000000000000000);
