+++ subgraphs/uniswap-forks/protocols/trader-joe/src/mappings/masterchef/rewardV3.ts
@@ -18,7 +18,7 @@ import { updateMasterChef } from "../../common/handlers/handleRewardV2";
 import {
   createMasterChefStakingPool,
   updateMasterChefTotalAllocation,
-} from "../../common/helpers";
+} from "../../../../../src/common/masterchef/helpers";
 
 // A deposit or stake for the pool specific MasterChef.
 export function handleDeposit(event: Deposit): void {
