+++ subgraphs/uniswap-forks/protocols/vvs-finance/src/common/handlers/handleReward.ts
@@ -8,18 +8,20 @@ import {
 } from "../../../../../generated/schema";
 import {
   BIGINT_ONE,
-  BIGINT_ZERO,
   INT_ZERO,
   MasterChef,
   RECENT_BLOCK_THRESHOLD,
 } from "../../../../../src/common/constants";
 import { getOrCreateToken } from "../../../../../src/common/getters";
 import { getRewardsPerDay } from "../../../../../src/common/rewards";
-import { getOrCreateMasterChef } from "../helpers";
 import {
   convertTokenToDecimal,
   roundToWholeNumber,
 } from "../../../../../src/common/utils/utils";
+import {
+  getOrCreateMasterChef,
+  getOrCreateMasterChefStakingPool,
+} from "../../../../../src/common/masterchef/helpers";
 
 // Called on both deposits and withdraws into the MasterApe/MasterChef pool.
 // Tracks staked LP tokens, and estimates the emissions of LP tokens for the liquidity pool associated with the staked LP.
@@ -33,8 +35,7 @@ export function handleReward(
   let masterChefPool = getOrCreateMasterChefStakingPool(
     event,
     MasterChef.MASTERCHEF,
-    pid,
-    poolContract
+    pid
   );
   let masterChef = getOrCreateMasterChef(event, MasterChef.MASTERCHEF);
 
@@ -137,32 +138,4 @@ export function handleReward(
   masterChef.save();
   rewardToken.save();
   pool.save();
-}
-
-// Create a MasterChefStaking pool using the MasterChef pid for id.
-function getOrCreateMasterChefStakingPool(
-  event: ethereum.Event,
-  masterChefType: string,
-  pid: BigInt,
-  poolContract: Craftsman
-): _MasterChefStakingPool {
-  let masterChefPool = _MasterChefStakingPool.load(
-    masterChefType + "-" + pid.toString()
-  );
-
-  // Create entity to track masterchef pool mappings
-  if (!masterChefPool) {
-    masterChefPool = new _MasterChefStakingPool(
-      masterChefType + "-" + pid.toString()
-    );
-
-    masterChefPool.multiplier = BIGINT_ONE;
-    masterChefPool.poolAllocPoint = BIGINT_ZERO;
-    masterChefPool.lastRewardBlock = event.block.number;
-    log.warning("MASTERCHEF POOL CREATED: " + pid.toString(), []);
-
-    masterChefPool.save();
-  }
-
-  return masterChefPool;
 }
