+++ subgraphs/uniswap-forks/subgraph.yaml
@@ -4,11 +4,11 @@ schema:
 dataSources:
   - kind: ethereum
     name: Factory
-    network: fantom
+    network: avalanche
     source:
-      address: "0x152ee697f2e276fa89e96742e9bb9ab1f2e61be3"
+      address: "0x9ad6c38be94206ca50bb0d90783181662f0cfa10"
       abi: Factory
-      startBlock: 3795376
+      startBlock: 2486392
     mapping:
       kind: ethereum/events
       apiVersion: 0.0.6
@@ -16,30 +16,30 @@ dataSources:
       entities: []
       abis:
         - name: Factory
-          file: ./abis/spookyswap/Factory.json
+          file: ./abis/trader-joe/Factory.json
         - name: TokenABI
-          file: ./abis/spookyswap/IERC20.json
+          file: ./abis/trader-joe/ERC20.json
       eventHandlers:
         - event: PairCreated(indexed address,indexed address,address,uint256)
           handler: handlePairCreated
       file: ./src/mappings/factory.ts
   - kind: ethereum
-    name: MasterChef
-    network: fantom
+    name: MasterChefV2
+    network: avalanche
     source:
-      address: "0x2b2929e785374c651a81a63878ab22742656dcdd"
-      abi: MasterChefSpookyswap
-      startBlock: 4267037
+      address: "0xd6a4F121CA35509aF06A0Be99093d08462f53052"
+      abi: MasterChefV2TraderJoe
+      startBlock: 2486444
     mapping:
       kind: ethereum/events
       apiVersion: 0.0.6
       language: wasm/assemblyscript
       entities: []
       abis:
-        - name: MasterChefSpookyswap
-          file: ./abis/spookyswap/MasterChefSpookyswap.json
+        - name: MasterChefV2TraderJoe
+          file: ./abis/trader-joe/MasterChefV2TraderJoe.json
         - name: TokenABI
-          file: ./abis/spookyswap/IERC20.json
+          file: ./abis/trader-joe/ERC20.json
       eventHandlers:
         - event: Deposit(indexed address,indexed uint256,uint256)
           handler: handleDeposit
@@ -47,40 +47,46 @@ dataSources:
           handler: handleWithdraw
         - event: EmergencyWithdraw(indexed address,indexed uint256,uint256)
           handler: handleEmergencyWithdraw
-      file: ./protocols/spookyswap/src/mappings/masterchef/reward.ts
-
+        - event: Add(indexed uint256,uint256,indexed address,indexed address)
+          handler: handleAdd
+        - event: Set(indexed uint256,uint256,indexed address,bool)
+          handler: handleSet
+        - event: UpdateEmissionRate(indexed address,uint256)
+          handler: handleUpdateEmissionRate
+      file: ./protocols/trader-joe/src/mappings/masterchef/rewardV2.ts
   - kind: ethereum
-    name: MasterChefV2
-    network: fantom
+    name: MasterChefV3
+    network: avalanche
     source:
-      address: "0x18b4f774fdc7bf685daeef66c2990b1ddd9ea6ad"
-      abi: MasterChefV2Spookyswap
-      startBlock: 34757343
+      address: "0x188bed1968b795d5c9022f6a0bb5931ac4c18f00"
+      abi: MasterChefV3TraderJoe
+      startBlock: 6034473
     mapping:
       kind: ethereum/events
       apiVersion: 0.0.6
       language: wasm/assemblyscript
-      entities: 
-        - _CircularBuffer
-        - RewardToken
+      entities: []
       abis:
-        - name: MasterChefV2Spookyswap
-          file: ./abis/spookyswap/MasterChefV2Spookyswap.json
+        - name: MasterChefV3TraderJoe
+          file: ./abis/trader-joe/MasterChefV3TraderJoe.json
         - name: TokenABI
-          file: ./abis/spookyswap/IERC20.json
+          file: ./abis/trader-joe/ERC20.json
       eventHandlers:
-        - event: Deposit(indexed address,indexed uint256,uint256,indexed address)
-          handler: handleDepositV2
-        - event: Withdraw(indexed address,indexed uint256,uint256,indexed address)
-          handler: handleWithdrawV2
-        - event: EmergencyWithdraw(indexed address,indexed uint256,uint256,indexed address)
-          handler: handleEmergencyWithdrawV2
-      file: ./protocols/spookyswap/src/mappings/masterchef/rewardV2.ts
- 
+        - event: Deposit(indexed address,indexed uint256,uint256)
+          handler: handleDeposit
+        - event: Withdraw(indexed address,indexed uint256,uint256)
+          handler: handleWithdraw
+        - event: EmergencyWithdraw(indexed address,indexed uint256,uint256)
+          handler: handleEmergencyWithdraw
+        - event: Add(indexed uint256,uint256,indexed address,indexed address)
+          handler: handleAdd
+        - event: Set(indexed uint256,uint256,indexed address,bool)
+          handler: handleSet
+      file: ./protocols/trader-joe/src/mappings/masterchef/rewardV3.ts
 templates:
   - kind: ethereum/contract
     name: Pair
-    network: fantom
+    network: avalanche
     source:
       abi: Pair
     mapping:
@@ -91,11 +97,11 @@ templates:
       entities: []
       abis:
         - name: Pair
-          file: ./abis/spookyswap/Pair.json
+          file: ./abis/spiritswap/Pair.json
         - name: Factory
-          file: ./abis/spookyswap/Factory.json
+          file: ./abis/spiritswap/Factory.json
         - name: TokenABI
-          file: ./abis/spookyswap/IERC20.json
+          file: ./abis/spiritswap/ERC20.json
       eventHandlers:
         - event: Mint(indexed address,uint256,uint256)
           handler: handleMint
@@ -107,31 +113,3 @@ templates:
           handler: handleTransfer
         - event: Sync(uint112,uint112)
           handler: handleSync
-  - kind: ethereum/contract
-    name: StakingRewards
-    network: celo
-    source:
-      abi: StakingRewards
-    mapping:
-      kind: ethereum/events
-      apiVersion: 0.0.6
-      language: wasm/assemblyscript
-      file: ./protocols/ubeswap/src/mappings/stakingrewards/handleReward.ts
-      entities: []
-      abis:
-        - name: StakingRewards
-          file: ./abis/ubeswap/StakingRewards.json
-        - name: PoolManager
-          file: ./abis/ubeswap/PoolManager.json
-        - name: TokenABI
-          file: ./abis/ubeswap/ERC20.json
-        - name: UniswapV2Pair
-          file: ./abis/ubeswap/UniswapV2Pair.json
-          
-      eventHandlers:
-        - event: Staked(indexed address,uint256)
-          handler: handleStaked
-        - event: Withdrawn(indexed address,uint256)
-          handler: handleWithdrawn
-        - event: RewardPaid(indexed address,uint256)
-          handler: handleRewardPaid
